Phase 3: frontend Vitest coverage + repo-wide fmt/clippy cleanup
Frontend now has five new .test.ts files covering the high-traffic
utilities and stores. Combined with the slideshowQueue +
thumbnailQueue suites Phase 1 migrated, vitest runs 61 tests in
~2 seconds:

- thumbnail.test.ts — thumbUrl path resolution: absolute drive-
  letter, UNC, forward-slash, relative-against-drive-root, missing
  inputs.

- stores/browseContext.test.ts — prev/next/position/extend/clear,
  including the "dedup ids already in the list" guarantee and
  null-on-edge behaviour.

- stores/selection.test.ts — toggle / set / range / clear, plus
  handleCellClick across plain / Ctrl-click / Meta-click /
  Shift-range / already-active-selection paths.

- stores/slideshow.test.ts — start / close / next / prev /
  toggleLoop / setInterval clamping / loadMore-on-end / loadMore
  dedup. Captures the singleton-state-leak gotcha in beforeEach
  so future tests don't trip over it.

- stores/devMode.test.ts — localStorage round-trip for the
  Settings dev-mode gate.

Vitest config now plugs in `@sveltejs/vite-plugin-svelte` so the
`$state` rune in `.svelte.ts` store files transpiles at test time.
jsdom added as a dev-dep so localStorage-backed tests work.

Repo-wide `cargo fmt` + `cargo clippy -D warnings` cleanup as a
side-effect of running `scripts/ci_local.sh ci`: two doc-list
indentation issues in `raw_preview.rs`, one unnecessary cast, one
overly-complex tuple type in `workflow_geocoding.rs`, plus the
usual rustfmt nits across the workspace. CI now passes end-to-end:

  cargo fmt --all --check
  cargo clippy --all-targets -D warnings
  cargo test (208 tests across all targets)
  npm run check + npm run test (61 vitest tests)

Wall-time for the full suite locally: ~10 s after the initial
build cache is warm.

+++ src/services/raw_preview.rs
@@ -29,12 +29,12 @@
 //!    - IFD chain via the NextIFD offset (last 4 bytes of each IFD)
 //!    - SubIFD chain via tag 0x014A
 //!    - EXIF / GPS / Interop SubIFDs via tags 0x8769 / 0x8825 / 0xA005
-//!    Each visit short-circuits if we've already seen that offset,
-//!    so a malformed file with a cycle can't hang us.
+//!      Each visit short-circuits if we've already seen that offset,
+//!      so a malformed file with a cycle can't hang us.
 //! 3. For each IFD, look for two embedded-JPEG idioms:
 //!    a. `JPEGInterchangeFormat` (0x0201) + length (0x0202)
 //!    b. `StripOffsets` (0x0111) + byte counts (0x0117) where
-//!       `Compression` (0x0103) is 6 (old-style JPEG) or 7 (new).
+//!    `Compression` (0x0103) is 6 (old-style JPEG) or 7 (new).
 //! 4. Sanity-bound each candidate against the file size, then return
 //!    the bytes of the largest one (or None if no candidate found).
 //!
@@ -61,10 +61,7 @@ use std::path::Path;
 /// decode this photo" rather than a hard error.
 pub fn extract_largest_preview(path: &Path) -> Result<Option<Vec<u8>>, String> {
     let file = File::open(path).map_err(|e| format!("open: {e}"))?;
-    let file_len = file
-        .metadata()
-        .map_err(|e| format!("stat: {e}"))?
-        .len();
+    let file_len = file.metadata().map_err(|e| format!("stat: {e}"))?.len();
     let mut reader = TiffReader::new(BufReader::new(file), file_len)?;
     let candidates = reader.find_candidates()?;
     let Some(best) = candidates.into_iter().max_by_key(|c| c.length) else {
@@ -361,8 +358,8 @@ impl TiffReader {
             return Err(format!("expected count=1, got {count}"));
         }
         match type_id {
-            3 => Ok((value & 0xFFFF) as u32), // inline SHORT
-            4 => Ok(value),                   // inline LONG
+            3 => Ok(value & 0xFFFF), // inline SHORT
+            4 => Ok(value),          // inline LONG
             _ => Err(format!("unsupported type {type_id}")),
         }
     }
